+++ Project 0/Assets/Script/Bomb/BombDamage/IDamagableFromBomb.cs
@@ -4,6 +4,6 @@ namespace Wepons.Bomb
 {
     public interface IDamagableFromBomb
     {
-        void TakeDamageFromBomb(float damage);
+        void TakeDamageFromBomb(int damage);
     }
 }
